Merged in jafingerhut/p4pktgen/maybe-good-enhancements (pull request #3)
Maybe good enhancements

Add support for boolean 'and' and 'or' operators, in addition to
'not'.

Add support for 'assign' primitive, which I believe is same in
behavior as 'modify_field'.

Allow tables with match_type 'lpm' as well as 'exact'.  As long as we
generate table entries for only one conditional path at a time, only a
single table entry in each table should be enough, and the match_type
should not matter much.

I have tried running some JSON files where the Concat(sym_key_elems)
gave an error because sym_key_elems contained only 1 element.  I think
perhaps a correct way to handle it when it has only 1 element is to
just use that first element.