Update representation of control_path and expected_path
Change control_path to be a list of tuples, where each tuple is
either:

+ A table and one of its actions, represented as the tuple
  (table_name, action_name)

+ A conditional node and whether the True or False branch is taken,
  represented as the tuple (conditional_node_name, True/False,
  source_filename, source_line_number, source_fragment_string).  The
  source information helps to map the output of simple_switch's log
  back to the node name in a nearly unambiguous way.

expected_path is also a list of the same kinds of tuples, with the
list of parse nodes at the beginning.

Including the True/False branch in the conditional nodes is expected
to make it easier to take an arbitrary length prefix of a control_path
list, and generate constraints from it.