Add demo program read-ingress-port.{p4,json}
This program correctly reads a field standard_metadata.ingress_port
that should be considered to be initialized when the parser begins,
but currently p4pktgen considers it uninitialized.

This field should be part of what the solver determines the contents
of, similar to the contents of the input packet.

There are likely to be a handful more similar fields that we discover
in the near future, so ideally it should be straightforward to add
more of them to a list somewhere in p4pktgen.