Update code for generating Graphviz graphs
It should probably have a new command line option to enable it, rather
than being on by default.  I haven't implemented that yet.

It has some code in there for coloring condition nodes red that
contain 'isValid' in the source fragment of the condition, with some
caveats mentioned in a comment in the source.

It also checks for actions that have add_header and/or remove_header
primitive operations within it, and colors those green, plus adds a
text label +<number_of_add_header_primitive_ops> and
-<number_of_remove_header_primitive_ops>