Fix some bugs in translator
+ min_bits_for_uint returned incorrect value for integers around
  2^48-1 or so, most likely due to floating point values to represent
  the difference between 48-epsilon and 48.0.  Changed implementation
  to use exact integer arithmetic.

+ Some constant expressions have type 'long', not 'int'.  Handled that
  like 'int'.

+ parser 'set' operations, and primitive 'assign' operators, were
sometimes putting in the context expressions for a field that were a
different number of bits wide than the declared width of the field in
the P4 program.  This was preventing finding solutions for IPv4 header
checksums, at least with the P4 program I had written to do it.

In general, it seems like a good idea to force the right-hand side
expression to be the width of the left-hand side destination field of
the assignment.

The P4_16 language spec requires assignments to be between equal-width
values, so as long as the P4_16 compiler is implemented correctly, it
should catch any attempts to do assignments between different bit
widths at compile time, never producing a JSON file for such programs.

I don't think this should cause issues for P4_14 programs.  It allows
assignments between values of different widths, with silent conversion
to the width of the left-hand side.  At least if you enable debugging
in p4pktgen, you get warnings when this conversion is happening.