Update 'small' incremental checksum program to one that takes about 10 mins
At least on a recent MacBookPro with a 2.2GHz Intel Core i7.  I
wouldn't recommend putting it in the list of short quick tests to be
done before check-in, but it shows what is near the limit of many
people's patience for verifying that two ways of calculating an IPv4
header checksum are equivalent.