Update P4_16 version of switch.p4 via retranslating from P4_14 with latest p4c
Also merge in the few P4_16 changes into the hand-edited
switch-p416-nohdrstack.p4.  While we are at this, replace the call to
random() in swithc-p416-nohdrstacks.p4 with assignment to the variable
from a new action parameter, so p4pktgen can control the random value
selected.

Also regenerate the JSON versions.