Rename P4 programs with issue #983 workaround as having it
For 2 of them, create versions that are nearly identical, but do not
have the workaround.

I have tested p4pktgen on the new
compiled_p4_programs/chksum-incremental1-small.json, compiled with the
latest p4c-bm2-ss compiler after issue #983 has been fixed
(hopefully).

It gives the same results as for the JSON produced before the bug was
fixed, on the P4 source program
p4_programs/chksum-incremental1-small-issue983-workaround.p4, which
worked around the bug in the P4_16 source code.