Add new automated test case exercising varbit extract and lookahead
Also add new command line argument --disable-packet-length-errors that
enables option that already exists in the code, and is the default
behavior when running all current pytest tests.

Add assert failure if we ever attempt to store two different
TestPathResult values for the same path in the results dict.  If this
ever happens, then I don't trust our pytest tests for such test cases,
and we should prevent that assert from occurring.

Miscellaneous updates to a few scripts in tools directory.