Support ternary operator (#12)
Tested with some programs I will add as examples soon.

When ternary operators are used in P4_16 source programs in a parser,
they are currently eliminated by transforming the parser into one with
more states, one that takes the true branch, the other taking the
false branch.  So it appears that at least at the moment, no separate
support for handling the ternary operator inside parser "set"
primitive operations is needed, but that might change in the future.