Update chksum-incremental-wrong-rfc1624-eqn2 programs
The P4 source program with issue983-workaround in its name was missing
a couple of workarounds for p4c issue #983, and thus sometimes getting
TEST_FAIL results.  No further bugs in p4pktgen or p4c were found
during testing these programs -- they confirm that with the additional
workarounds, it is working fine, and p4c issue #983 seems to be fixed
for the version of the P4 program without the workarounds for p4c
issue #983.