Delay error for multiple condition nodes with same source info (#66)
Before it was done while reading the bmv2 JSON file.  With this
change, this is allowed, and the error is only printed if we are
running simple_switch, and trying to map the source info for a
condition evaluation in the console log back to a unique node in the
control flow graph.

This enables running p4pktgen _without_ the --run-simple-switch
command line option, and still being able to generate test cases for
P4 programs that have such bmv2 JSON files.

I have found several examples of such programs.  For example, if you
have a control that is called from the ingress control multiple times,
and the sub-control has an 'if' statement in it, then while p4c-bm2-ss
in-lines these calls to the sub-control, it can create multiple nodes
in the control flow that all have the same source info, i.e. the same
file name and line number, of the line in the sub-control being
inlined.

This change also includes optional parsing of a node name in the
simple_switch console log.  Such a node name is not printed now, but I
have submitted a small pull request to p4lang/behavioral-model that
would add this.  It is a small enough change that it may be accepted.
If so, then there source info will no longer be needed to determine
the unique node name associated with the condition evaluation.