Add example program exercising many header stack operations
This is a program that I wrote, along with several example-based
packet tests in an STF file, in this Github PR:

    https://github.com/p4lang/p4c/pull/1129

It has at least 23^3 paths in it as written, so way more than you want
to exercise with p4pktgen.  We could pretty easily comment out or
delete a bunch of the cases to cut it down.  Also perhaps call the
sub-control 2 times instead of 3 to change exponent from 3 to 2.  Also
perhaps replace the 5 if statements that fill in the output header
bits with a single expression with shifting and bitwise ORing to avoid
the ifs, then follow that with a table search on a key containing
those 5 bits (and the rest of the 8 bits it is in, if you want).  That
would force p4pktgen to put its predicted value for those 5 bits into
a table search key in order to match.