Add switch-p416-nohashrandom.p4 and BMv2 JSON file for it
The P4_16 source code is nearly identical to switch-p416.p4, except it
comments out uses of the action_profile and action_selector externs,
and also calls to hash() and random(), replacing them with assignments
to new action parameters, so that p4pktgen can solve for their
resulting values.

Compiled using latest version of p4c-bm2-ss as of today.