Update most test programs to use latest p4c as of 2019-Jun-07
Update 7 out of 10 of the test programs to use the latest version of
p4c as of 2019-Jun-07 for their compiled BMv2 JSON files.

The following 3 still use the latest version of p4c as of 2019-Jan-06,
as they did before this commit, because starting with a commit on
2019-May-04, p4c will in some cases compile 'exit' statements in P4_16
source code into using the "exit" primitive action in BMv2 JSON files,
and p4pktgen does not handle those yet.  p4pktgen issue number 80 is
tracking this.

demo1b.p4
demo9b.p4
checksum-ipv4-with-options.p4