Enforce packet-length limits when solving.
The previous approach of filtering out packets that were not at least of the
minimum length was problematic for several reasons, but in particular it would
allow the solver to return a zero-length packet, which trips assertions in
unrelated places.
The maximum length limit was not previously applied, which led to very
large packets being generated for programs with empty parser graphs.