Add option to load extern backends.
This option is used to pass in python models of extern backends.  The
models should use z3 directly to generate expressions that represent the
action they are modelling.  An example model is included in the tests of
this feature.
This commit ensures that backends are available, the next commit enables
programs to actually use them.