Removed header field size lookup, added comment to fields.
The get size function is currently unused.  As written, it also risks
causing an uninitialised read if called.  It's easier and safer to just
do this by looking at the contents of the fields member in Context.
Added comment to the fields dict to make it clearer what it contains.