Calculate TestPathResult before building test case.
The Solver needs to know the path result, which was previously done
during the test case build.  We want to move test case building out of
the Solver, and it's easy enough to calculate the result without
building the test case.  We check that the build and test results are
still equivalent just to be safe.