Added round-robin-parser-paths option.
Using this option causes parser paths to be round robined, finding a
test case for the current one each time.  This can be used when
generating a small number of paths for a large program to achieve better
coverage of parser paths than if the iteration were linear.
Note that as state for each parser path must be stored independently,
using this option on programs with a large number of parser paths may
consume a large amount of memory.