Updated broken parallel-parser test to check existing behavior.
The previous implementation of the test was fundamentally broken, but
looked to be checking for an issue that was solved in p4pktgen itself
some time ago.  Solving the same issue in the tests is a bit trickier,
but can be worked around by representing the results as a list.