Added collapse-parser-paths option.
The option swaps parallel parser paths with a ParserCompositeTransition
that represents all of the paths at the preprocess stage of the visitor.
The graph itself is unaltered, the new transition contains all the graph
edges, but is not part of the graph itself.  Test cases generated on
paths including this transition may traverse any of the parallel
transitions at that point in the parser graph.