for-upstream/various-refactor (#106)
* Cleanup: Fixed stale usage of HLIR_Parser_Transition class name.

ParserTransition has used this class name for some time, but still had
one remaining reference to the old class name.

* Cleanup: Removed unused/unnecessary variables and imports from main.

* Purged non-functional p4 input code.

Not functional and removing it makes further refactoring a bit simpler.

* Refactored common graph building code into P4_Top.

Packaging graphs into P4_Top means there are fewer things to keep track
of in the higher level function.

* Refactored graphviz visualization codepath.

Moving this into it's own utility module makes the main module easier to
work on.

* Refactored printout of parser paths.

Moving into a separate function makes the generate_test_cases function
easier to work on.

* Moved test config setting function into test code.

The config used for tests has more to do with the test framework than
the main library, so moved it there for simplicity.

* Made output file path part of config.

It doesn't need to be configurable yet, but it's better to standardise
the output names of json and pcap files so that they are more clearly
linked.

* Use debug argument in Config rather than local variable.

* Un-nested classes in p4_hlir.

This should not introduce any functional change, it just moves all
classes in p4_hlir from being nested to all being defined at the module
level.  This is to allow instances of the HLIR to be pickled, something
that is disallowed by nesting classes.

* Switched ops enum to class type, removed unused enum.

Functionally defined enums cannot be pickled without specifying qualname.

* Restrict the Z3 solver to QF_UFBV.

The default logic is larger than necessary.  Restricting it in this way gives
a performance improvement of 30-40% according to measurements.

Co-authored-by: Gregor Dick <gregord@xilinx.com>
Co-authored-by: Andy Fingerhut <andy_fingerhut@alum.wustl.edu>