Separate packet vars for each extractions
Previously, p4pktgen supported two modes for modelling extractions in the
parser: either the packet was modelled by a single wide BitVec Z3 varible, or,
in "hybrid mode" as an optimisation, as a series of narrow variables modelling
fixed-length extractions, concatenated with the wide vector.

Hybrid mode had a bug: it used z3.is_const() to decide whether a given
extraction was fixed-length and therefore eligible for hybridisation, but the
semantics of that function do not match our requirements.

This patch re-implements the packet-construction logic in order to avoid this
problem, such that each extraction is always modelled as its own Z3 variable,
but with the necessary handling for variable-length reads and lookaheads.