Refactored record_result logic into two functions.
Splits what was previously record_result into two functions:
record_path_result, whether to store the path and result in the dict
that is returned to tests.
record_test_case, whether to write a test case for the path.

The only functional change is that NO_PACKET_FOUND paths no longer have
test cases written.  This seems better in general, and is required for
table consolidation as we can only consolidate paths that have solutions.