Added option to solve paths while consolidating table configs.
The -ct option causes p4pktgen to pass solved paths into a new solver
that attempts to consolidate table configs between the paths and then
solve all the paths again together.  If a solution cannot be found then
it flushes the existing solved set of paths and starts a new set.
The -ct option can accept a number, limiting the number of paths to
group together.  Performance can suffer when attempting to solve large
numbers of complex paths at once.