Moved ConsolidatedSolver base class to a push/pop model.
This helps solve the problem of tracking constraints handled in the next
commit.  Doing something similar with the variables managed by
TableConsolidatedSolver is possible, but not quite as easy as they're
not simple lists.  It's also overkill as we'll never need to pop back
more than one state from the tail.